+++ src/Abm Turno/ModificacionTurno.cs
@@ -24,7 +24,7 @@ namespace UberFrba.Abm_Turno
         {
             if (verificarDatosTurno(txtDescripcion.Text, txtValorKm.Text, txtPrecioBase.Text))
             {
-                Turno nuevoTurno = new Turno(idTurno, Int32.Parse(sacarHoraReal(cmbInicio.Text)), Int32.Parse(sacarHoraReal(cmbFinal.Text)), txtDescripcion.Text, double.Parse(txtValorKm.Text), double.Parse(txtPrecioBase.Text), ckbHabilitado.Checked);
+                Turno nuevoTurno = new Turno(idTurno, Int32.Parse(sacarHoraReal(cmbInicio.Text)), Int32.Parse(sacarHoraReal(cmbFinal.Text)), txtDescripcion.Text, decimal.Parse(txtValorKm.Text), decimal.Parse(txtPrecioBase.Text), ckbHabilitado.Checked);
                 string response=SQLTurno.modificarTurno(nuevoTurno);
                 this.DialogResult = DialogResult.OK;
                 MessageBox.Show(response);
